Issue #793 Fix OpenSSL Only patch OpenSSL if it is present

+++ assets/rakelib/ruboto.stdlib.rake
@@ -262,14 +262,16 @@ def cleanup_jars
         # FIXME(uwe): Use our own patched version of OpenSSL until jruby-openssl 0.9.16 is released
         # FIXME(uwe): Remove before Ruboto release!
         # FIXME(uwe): Remove after jruby-openssl 0.9.16 is added to the jruby-1_7 branch
-        puts '*' * 80
-        puts "Patching OpenSSL"
-        FileUtils.rm 'org/jruby/ext/openssl/OpenSSL.class'
-        FileUtils.mkdir_p '../../../src/org/jruby/ext/openssl'
-        require 'ruboto'
-        FileUtils.cp "#{Ruboto::ASSETS}/src/org/jruby/ext/openssl/OpenSSL.java",
-            '../../../src/org/jruby/ext/openssl/OpenSSL.java'
-        puts '*' * 80
+        if File.exists? 'org/jruby/ext/openssl/OpenSSL.class'
+          puts '*' * 80
+          puts 'Patching OpenSSL'
+          FileUtils.rm 'org/jruby/ext/openssl/OpenSSL.class'
+          FileUtils.mkdir_p '../../../src/org/jruby/ext/openssl'
+          require 'ruboto'
+          FileUtils.cp "#{Ruboto::ASSETS}/src/org/jruby/ext/openssl/OpenSSL.java",
+              '../../../src/org/jruby/ext/openssl/OpenSSL.java'
+          puts '*' * 80
+        end
         # EMXIF
         jar_load_code = <<-END_CODE
           require 'jruby'
